requirements.txt
streamlit
google-generativeai
Pillow
flask
python-dotenv
gunicorn
